+++ src/com/jxhh/goods/GoodsStorageOnlineList.java
@@ -60,7 +60,7 @@ public class GoodsStorageOnlineList {
     private String unit;
     private Long total_stock;
     private Integer activity_price;
-    private Integer created_time;
+    private Long created_time;
     private String dis_category_id;
     private Integer source;
     private String cover;
@@ -74,7 +74,7 @@ public class GoodsStorageOnlineList {
     private Integer real_sale;
     private Long shop_id;
     private Long third_id;
-    private Integer updated_time;
+    private Long updated_time;
     private Integer guide_price;
     private Integer sale_price;
     private Integer cost_price;
@@ -167,11 +167,11 @@ public class GoodsStorageOnlineList {
         this.activity_price = activity_price;
     }
 
-    public Integer getCreated_time() {
+    public Long getCreated_time() {
         return created_time;
     }
 
-    public void setCreated_time(Integer created_time) {
+    public void setCreated_time(Long created_time) {
         this.created_time = created_time;
     }
 
@@ -279,11 +279,11 @@ public class GoodsStorageOnlineList {
         this.third_id = third_id;
     }
 
-    public Integer getUpdated_time() {
+    public Long getUpdated_time() {
         return updated_time;
     }
 
-    public void setUpdated_time(Integer updated_time) {
+    public void setUpdated_time(Long updated_time) {
         this.updated_time = updated_time;
     }
